Seperate components from one file
하나의 파일을 컴퍼넌트 별로 나눴습니다.
