+++ src/AddButton.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export default function AddButton({ onClickAddTodo }) {
+  return (
+    <button
+      type="submit"
+      onClick={onClickAddTodo}
+    >
+      추가
+    </button>
+  );
+}
