+++ src/Todo.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default function Todo({ title, id, onClick }) {
+  return (
+    <li key={id.toString()}>
+      {title}
+      <button type="button" onClick={() => onClick(id)}>완료</button>
+    </li>
+  );
+}
